No default schema, moving icons
