+++ src/lib/ui/icons/ErrorIcon.svelte
