+++ src/lib/ui/icons/InvoiceIcon.svelte
