+++ src/lib/ui/icons/LightbulbIcon.svelte
