+++ src/lib/ui/icons/MessageIcon.svelte
