+++ src/lib/ui/icons/SuccessIcon.svelte
