+++ src/lib/ui/icons/WarningIcon.svelte
