+++ src/templates/es/invoice-freelance.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://gobl.org/draft-0/bill/invoice",
-  "code": "SAMPLE-001",
-  "currency": "EUR",
+  "series": "SAMPLE",
+  "code": "001",
   "issue_date": "2022-02-01",
   "supplier": {
     "tax_id": {
