+++ src/templates/es/invoice.json
@@ -1,8 +1,9 @@
 {
   "$schema": "https://gobl.org/draft-0/bill/invoice",
-  "code": "SAMPLE-001",
+  "series": "S2023",
+  "code": "001",
   "currency": "EUR",
-  "issue_date": "2022-02-01",
+  "issue_date": "2023-08-01",
   "supplier": {
     "tax_id": {
       "country": "ES",
@@ -56,13 +57,13 @@
     {
       "quantity": "1",
       "item": {
-        "name": "Financial service",
+        "name": "Promotional cookies",
         "price": "10.00"
       },
       "taxes": [
         {
           "cat": "VAT",
-          "rate": "zero"
+          "rate": "reduced"
         }
       ]
     }
