+++ src/templates/nl/invoice.json
@@ -1,6 +1,7 @@
 {
   "$schema": "https://gobl.org/draft-0/bill/invoice",
-  "code": "HS2022071201",
+  "series": "HS2022",
+  "code": "071201",
   "currency": "EUR",
   "issue_date": "2022-07-12",
   "supplier": {
